+++ src/rex.instrument/src/rex/instrument/interface/assessment.py
@@ -325,7 +325,7 @@ class Assessment(Extension, Comparable, Displayable, Dictable):
         return data
 
     @classmethod
-    def get_by_uid(cls, uid):
+    def get_by_uid(cls, uid, user=None):
         """
         Retrieves an Assessment from the datastore using its UID.
 
@@ -333,6 +333,8 @@ class Assessment(Extension, Comparable, Displayable, Dictable):
 
         :param uid: the UID of the Assessment to retrieve
         :type uid: string
+        :param user: the User who should have access to the desired Assessment
+        :type user: User
         :raises:
             DataStoreError if there was an error reading from the datastore
         :returns:
@@ -343,7 +345,7 @@ class Assessment(Extension, Comparable, Displayable, Dictable):
         raise NotImplementedError()
 
     @classmethod
-    def find(cls, offset=0, limit=100, **search_criteria):
+    def find(cls, offset=0, limit=100, user=None, **search_criteria):
         """
         Returns Asessments that match the specified criteria.
 
@@ -363,6 +365,8 @@ class Assessment(Extension, Comparable, Displayable, Dictable):
             the maximum number of Assessments to return (useful for pagination
             purposes)
         :type limit: int
+        :param user: the User who should have access to the desired Assessments
+        :type user: User
         :raises:
             DataStoreError if there was an error reading from the datastore
         :rtype: list of Assessments