+++ src/rex.instrument/src/rex/instrument/interface/draftinstrumentversion.py
@@ -53,7 +53,7 @@ class DraftInstrumentVersion(Extension, Comparable, Displayable, Dictable):
         iv_impl.validate_definition(definition)
 
     @classmethod
-    def get_by_uid(cls, uid):
+    def get_by_uid(cls, uid, user=None):
         """
         Retrieves a DraftInstrumentVersion from the datastore using its UID.
 
@@ -61,6 +61,10 @@ class DraftInstrumentVersion(Extension, Comparable, Displayable, Dictable):
 
         :param uid: the UID of the DraftInstrumentVersion to retrieve
         :type uid: string
+        :param user:
+            the User who should have access to the desired
+            DraftInstrumentVersion
+        :type user: User
         :raises:
             DataStoreError if there was an error reading from the datastore
         :returns:
@@ -71,7 +75,7 @@ class DraftInstrumentVersion(Extension, Comparable, Displayable, Dictable):
         raise NotImplementedError()
 
     @classmethod
-    def find(cls, offset=0, limit=100, **search_criteria):
+    def find(cls, offset=0, limit=100, user=None, **search_criteria):
         """
         Returns DraftInstrumentVersions that match the specified criteria.
 
@@ -90,6 +94,10 @@ class DraftInstrumentVersion(Extension, Comparable, Displayable, Dictable):
             the maximum number of DraftInstrumentVersions to return (useful for
             pagination purposes)
         :type limit: int
+        :param user:
+            the User who should have access to the desired
+            DraftInstrumentVersions
+        :type user: User
         :raises:
             DataStoreError if there was an error reading from the datastore
         :rtype: list of DraftInstrumentVersions