+++ src/rex.instrument/src/rex/instrument/interface/subject.py
@@ -21,7 +21,7 @@ class Subject(Extension, Comparable, Displayable, Dictable):
     """
 
     @classmethod
-    def get_by_uid(cls, uid):
+    def get_by_uid(cls, uid, user=None):
         """
         Retrieves a Subject from the datastore using its UID.
 
@@ -29,6 +29,8 @@ class Subject(Extension, Comparable, Displayable, Dictable):
 
         :param uid: the UID of the Subject to retrieve
         :type uid: string
+        :param user: the User who should have access to the desired Subject
+        :type user: User
         :raises:
             DataStoreError if there was an error reading from the datastore
         :returns:
@@ -39,7 +41,7 @@ class Subject(Extension, Comparable, Displayable, Dictable):
         raise NotImplementedError()
 
     @classmethod
-    def find(cls, offset=0, limit=100, **search_criteria):
+    def find(cls, offset=0, limit=100, user=None, **search_criteria):
         """
         Returns Subjects that match the specified criteria.
 
@@ -57,6 +59,8 @@ class Subject(Extension, Comparable, Displayable, Dictable):
             the maximum number of Subjects to return (useful for pagination
             purposes)
         :type limit: int
+        :param user: the User who should have access to the desired Subjects
+        :type user: User
         :raises:
             DataStoreError if there was an error reading from the datastore
         :rtype: list of Subjects